Refactor CDP Data Service client and enhance workload API request handling
