+++ plugins/module_utils/cdp_df.py
@@ -18,263 +18,18 @@
 A REST client for the Cloudera on Cloud Platform (CDP) DataFlow API
 """
 
-import json
 import time
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, List, Optional, Tuple
 
-from ansible.module_utils.urls import fetch_url
 from ansible_collections.cloudera.cloud.plugins.module_utils.cdp_client import (
     CdpClient,
     CdpError,
 )
 from ansible_collections.cloudera.cloud.plugins.module_utils.cdp_df_client import (
     CdpDfApiClient,
+    DataServiceClient,
 )
 from ansible_collections.cloudera.cloud.plugins.module_utils.cdp_iam import CdpIamClient
-
-
-class DataServiceClient(CdpClient):
-    """
-    CDP Data Service Workload client using Bearer token authentication.
-
-    Unlike the control plane (which authenticates with Ed25519 signatures),
-    data service workload APIs authenticate with a JWT Bearer token and rely
-    on an XSRF cookie for mutating requests. The token and base URL are
-    discovered through the control plane (e.g. IAM ``generateWorkloadAuthToken``
-    or ``initiateDeployment``) and supplied to this client.
-
-    It is currently used for the DataFlow workload APIs (``/dfx/api/...``) but
-    is service-agnostic and can back other data services that share the same
-    auth scheme.
-
-    Responsibilities:
-        - Bearer token authentication
-        - XSRF/CSRF token handling via automatic cookie capture
-        - Retry with exponential backoff on transient failures
-    """
-
-    def __init__(
-        self,
-        module,
-        base_url: str,
-        access_token: str,
-        timeout_seconds: int = 60,
-        default_page_size: int = 100,
-    ):
-        """
-        Initialize the workload client.
-
-        Args:
-            module: AnsibleModule instance (used for ``fetch_url`` and TLS config)
-            base_url: Base URL of the workload API
-            access_token: JWT Bearer token from generateWorkloadAuthToken
-            timeout_seconds: Per-request HTTP timeout in seconds
-            default_page_size: Default page size for paginated requests
-        """
-        super().__init__(default_page_size=default_page_size)
-        self.module = module
-        self.base_url = base_url.rstrip("/")
-        self.access_token = access_token
-        self.timeout_seconds = timeout_seconds
-        self.cookies: Dict[str, str] = {}
-        self.headers = {
-            "Content-Type": "application/json",
-            "Accept": "application/json",
-            "Authorization": f"Bearer {self.access_token}",
-        }
-
-    def _url(self, path: str) -> str:
-        """Construct the full request URL from the base URL and a path."""
-        return f"{self.base_url}/{path.lstrip('/')}"
-
-    def _make_request(
-        self,
-        method: str,
-        path: str,
-        params: Optional[Dict[str, Any]] = None,
-        data: Optional[Union[Dict[str, Any], List[Any]]] = None,
-        json_data: Optional[Union[Dict[str, Any], List[Any]]] = None,
-        max_retries: int = 3,
-        squelch: Dict[int, Any] = {},
-    ) -> Any:
-        """
-        Make a workload API request with cookie handling and retry logic.
-
-        Args:
-            method: HTTP method (GET, POST, PUT, DELETE)
-            path: API path relative to the base URL
-            params: Query parameters appended to the URL
-            data: Request body serialized as JSON
-            json_data: Request body serialized as JSON (takes precedence over data)
-            max_retries: Maximum number of attempts before failing
-            squelch: Map of status codes to values returned instead of raising
-
-        Returns:
-            Parsed JSON response, or None for 204 responses
-
-        Raises:
-            CdpError: On non-squelched HTTP errors or exhausted retries
-        """
-        url = self._url(path)
-        headers = self.headers.copy()
-
-        if params:
-            query_params = []
-            for key, value in params.items():
-                if isinstance(value, list):
-                    query_params.extend(f"{key}={item}" for item in value)
-                else:
-                    query_params.append(f"{key}={value}")
-            url = f"{url}?{'&'.join(query_params)}"
-
-        if self.cookies:
-            headers["Cookie"] = "; ".join(
-                [f"{k}={v}" for k, v in self.cookies.items()]
-            )
-            if "XSRF-TOKEN" in self.cookies:
-                headers["X-XSRF-TOKEN"] = self.cookies["XSRF-TOKEN"]
-
-        body = None
-        if json_data is not None:
-            body = json.dumps(json_data)
-        elif data is not None:
-            body = json.dumps(data)
-
-        original_validate_certs = self.module.params.get("validate_certs")
-        self.module.params["validate_certs"] = self.module.params.get(
-            "endpoint_tls", True
-        )
-
-        last_error = None
-        try:
-            for attempt in range(max_retries):
-                try:
-                    resp, info = fetch_url(
-                        self.module,
-                        url,
-                        data=body,
-                        headers=headers,
-                        method=method,
-                        timeout=self.timeout_seconds,
-                    )
-
-                    status_code = info.get("status", -1)
-
-                    set_cookie = info.get("set-cookie")
-                    if set_cookie:
-                        cookie_headers = (
-                            [set_cookie]
-                            if isinstance(set_cookie, str)
-                            else set_cookie
-                        )
-                        for cookie_header in cookie_headers:
-                            if "=" in cookie_header:
-                                cookie_part = cookie_header.split(";")[0]
-                                name, value = cookie_part.split("=", 1)
-                                self.cookies[name.strip()] = value.strip()
-
-                    if status_code in squelch:
-                        return squelch[status_code]
-
-                    if status_code == -1:
-                        error_msg = info.get("msg", "Connection error")
-                        if attempt < max_retries - 1:
-                            time.sleep(min(0.5 * (2**attempt), 5))
-                            last_error = CdpError(
-                                f"{error_msg} for {url}", status=status_code
-                            )
-                            continue
-                        raise CdpError(f"{error_msg} for {url}", status=status_code)
-
-                    if 200 <= status_code < 300:
-                        if status_code == 204:
-                            return None
-                        if resp:
-                            response_text = resp.read().decode("utf-8")
-                            if response_text:
-                                try:
-                                    return json.loads(response_text)
-                                except json.JSONDecodeError:
-                                    return {"response": response_text}
-                            return {}
-                        return {}
-
-                    error_message = f"HTTP {status_code} Error"
-                    try:
-                        error_body = info.get("body")
-                        if error_body:
-                            error_data = json.loads(error_body)
-                            if "message" in error_data:
-                                error_message = error_data["message"]
-                            elif "error" in error_data:
-                                error_message = error_data["error"]
-                            elif "errorMessages" in error_data:
-                                msgs = error_data["errorMessages"]
-                                error_message = (
-                                    " ".join(msgs) if isinstance(msgs, list) else str(msgs)
-                                )
-                            else:
-                                error_message = error_data.get(
-                                    "errorMessage", "Unknown error"
-                                )
-                        else:
-                            error_message = info.get("msg", "Unknown error")
-                    except Exception:
-                        error_message = info.get("msg", "Unknown error")
-
-                    if status_code >= 500 or status_code in [408, 429]:
-                        if attempt < max_retries - 1:
-                            time.sleep(min(0.5 * (2**attempt), 5))
-                            last_error = CdpError(
-                                f"{error_message} for {url}", status=status_code
-                            )
-                            continue
-
-                    raise CdpError(f"{error_message} [{status_code}] for {url}")
-
-                except CdpError:
-                    raise
-                except Exception as e:
-                    if attempt < max_retries - 1:
-                        time.sleep(min(0.5 * (2**attempt), 5))
-                        last_error = CdpError(f"Connection error for {url}: {str(e)}")
-                        continue
-                    raise CdpError(
-                        f"Request failed after {max_retries} attempts for {url}: {str(e)}"
-                    )
-
-            if last_error:
-                raise last_error
-            raise CdpError(f"Request failed for {url}")
-        finally:
-            if original_validate_certs is not None:
-                self.module.params["validate_certs"] = original_validate_certs
-
-    def get(self, path: str, params: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
-        return self._make_request("GET", path, params=params)
-
-    def post(
-        self,
-        path: str,
-        data: Optional[Dict[str, Any]] = None,
-        json_data: Optional[Dict[str, Any]] = None,
-        squelch: Dict[int, Any] = {},
-    ) -> Dict[str, Any]:
-        return self._make_request("POST", path, data=data, json_data=json_data, squelch=squelch)
-
-    def put(
-        self,
-        path: str,
-        data: Optional[Dict[str, Any]] = None,
-        json_data: Optional[Dict[str, Any]] = None,
-        squelch: Dict[int, Any] = {},
-    ) -> Dict[str, Any]:
-        """Execute an HTTP PUT request."""
-        return self._make_request("PUT", path, data=data, json_data=json_data, squelch=squelch)
-
-    def delete(self, path: str, squelch: Dict[int, Any] = {}) -> Dict[str, Any]:
-        """Execute an HTTP DELETE request."""
-        return self._make_request("DELETE", path, squelch=squelch)
 
 
 class DataFlowModule:
