+++ plugins/module_utils/cdp_df_client.py
@@ -27,9 +27,114 @@ from ansible.module_utils.urls import fetch_url
 
 from ansible_collections.cloudera.cloud.plugins.module_utils.cdp_client import (
     AnsibleCdpClient,
+    CdpClient,
     CdpError,
     make_signature_header,
 )
+
+
+# ============================================================================
+# HTTP request/response helpers for workload (Bearer-token) transport
+#
+# These hold the request-building and response-parsing logic used by
+# WorkloadTransport below. They are kept local to this DataFlow-specific module
+# rather than shared from cdp_client.py, which signs requests with Ed25519.
+# ============================================================================
+
+
+def build_query_string(url: str, params: Optional[Dict[str, Any]]) -> str:
+    """
+    Append query parameters to a URL.
+
+    List values expand into repeated ``key=item`` pairs. Returns the URL
+    unchanged when ``params`` is empty.
+    """
+    if not params:
+        return url
+
+    query_params: List[str] = []
+    for key, value in params.items():
+        if isinstance(value, list):
+            query_params.extend(f"{key}={item}" for item in value)
+        else:
+            query_params.append(f"{key}={value}")
+    return f"{url}?{'&'.join(query_params)}"
+
+
+def serialize_request_body(
+    data: Optional[Union[Dict[str, Any], List[Any]]] = None,
+    json_data: Optional[Union[Dict[str, Any], List[Any]]] = None,
+) -> Optional[str]:
+    """
+    Serialize a request body to a JSON string.
+
+    ``json_data`` takes precedence over ``data``. Returns None when neither is
+    provided.
+    """
+    if json_data is not None:
+        return json.dumps(json_data)
+    if data is not None:
+        return json.dumps(data)
+    return None
+
+
+def parse_success_response(resp: Any, status_code: int) -> Any:
+    """
+    Parse a 2xx HTTP response.
+
+    Returns None for 204, the decoded JSON object for a JSON body, a
+    ``{"response": <text>}`` wrapper for a non-JSON body, or an empty dict when
+    there is no body.
+    """
+    if status_code == 204:
+        return None
+    if resp:
+        response_text = resp.read().decode("utf-8")
+        if response_text:
+            try:
+                return json.loads(response_text)
+            except json.JSONDecodeError:
+                return {"response": response_text}
+        return {}
+    return {}
+
+
+def parse_error_message(info: Dict[str, Any], status_code: int) -> str:
+    """
+    Extract a human-readable error message from a fetch_url ``info`` dict.
+
+    Looks in the response body for ``message`` / ``error`` / ``errorMessages``
+    / ``errorMessage`` keys, falling back to ``info['msg']``.
+    """
+    error_message = f"HTTP {status_code} Error"
+    try:
+        error_body = info.get("body")
+        if error_body:
+            error_data = json.loads(error_body)
+            if "message" in error_data:
+                error_message = error_data["message"]
+            elif "error" in error_data:
+                error_message = error_data["error"]
+            elif "errorMessages" in error_data:
+                msgs = error_data["errorMessages"]
+                error_message = " ".join(msgs) if isinstance(msgs, list) else str(msgs)
+            else:
+                error_message = error_data.get("errorMessage", "Unknown error")
+        else:
+            error_message = info.get("msg", "Unknown error")
+    except Exception:
+        error_message = info.get("msg", "Unknown error")
+    return error_message
+
+
+def is_retryable_status(status_code: int) -> bool:
+    """Return True for status codes worth retrying (5xx, 408, 429)."""
+    return status_code >= 500 or status_code in (408, 429)
+
+
+def compute_backoff(attempt: int) -> float:
+    """Exponential backoff with a 5s ceiling: 0.5s, 1s, 2s, 4s, 5s, ..."""
+    return min(0.5 * (2**attempt), 5)
 
 
 class CdpDfApiClient(AnsibleCdpClient):
@@ -172,3 +277,228 @@ class CdpDfApiClient(AnsibleCdpClient):
             data=redirect_body,
             timeout=self.timeout,
         )
+
+
+class WorkloadTransport(CdpClient):
+    """
+    HTTP transport for CDP data-service workload APIs.
+
+    Workload APIs (e.g. DataFlow ``/dfx/api/...``) authenticate with a JWT
+    Bearer token plus an XSRF cookie, unlike the control plane which signs
+    requests with Ed25519. This base owns the request/retry/parse machinery so
+    concrete workload clients only declare their auth headers.
+
+    Responsibilities:
+        - XSRF/CSRF cookie capture and replay
+        - Retry with exponential backoff on transient failures
+        - Raises CdpError on failure (it does not call ``module.fail_json``)
+
+    Subclasses must set ``self.module``, ``self.base_url``, ``self.timeout``,
+    ``self.cookies`` and ``self.headers`` in their ``__init__``.
+    """
+
+    def _url(self, path: str) -> str:
+        """Construct the full request URL from the base URL and a path."""
+        return f"{self.base_url}/{path.lstrip('/')}"
+
+    def _request_headers(self) -> Dict[str, str]:
+        """Return per-request headers: base headers plus any captured cookies."""
+        headers = self.headers.copy()
+        if self.cookies:
+            headers["Cookie"] = "; ".join(
+                [f"{k}={v}" for k, v in self.cookies.items()]
+            )
+            if "XSRF-TOKEN" in self.cookies:
+                headers["X-XSRF-TOKEN"] = self.cookies["XSRF-TOKEN"]
+        return headers
+
+    def _capture_cookies(self, info: Dict[str, Any]) -> None:
+        """
+        Capture cookies from a response into ``self.cookies``.
+
+        Workload APIs hand back an XSRF token via ``Set-Cookie`` on the first
+        request; subsequent mutating requests must echo it back.
+        """
+        set_cookie = info.get("set-cookie")
+        if not set_cookie:
+            return
+
+        cookie_headers = [set_cookie] if isinstance(set_cookie, str) else set_cookie
+        for cookie_header in cookie_headers:
+            if "=" in cookie_header:
+                cookie_part = cookie_header.split(";")[0]
+                name, value = cookie_part.split("=", 1)
+                self.cookies[name.strip()] = value.strip()
+
+    def _make_request(
+        self,
+        method: str,
+        path: str,
+        params: Optional[Dict[str, Any]] = None,
+        data: Optional[Union[Dict[str, Any], List[Any]]] = None,
+        json_data: Optional[Union[Dict[str, Any], List[Any]]] = None,
+        max_retries: int = 3,
+        squelch: Dict[int, Any] = {},
+    ) -> Any:
+        """
+        Make a workload API request with cookie handling and retry logic.
+
+        Args:
+            method: HTTP method (GET, POST, PUT, DELETE)
+            path: API path relative to the base URL
+            params: Query parameters appended to the URL
+            data: Request body serialized as JSON
+            json_data: Request body serialized as JSON (takes precedence over data)
+            max_retries: Maximum number of attempts before failing
+            squelch: Map of status codes to values returned instead of raising
+
+        Returns:
+            Parsed JSON response, or None for 204 responses
+
+        Raises:
+            CdpError: On non-squelched HTTP errors or exhausted retries
+        """
+        url = build_query_string(self._url(path), params)
+        headers = self._request_headers()
+        body = serialize_request_body(data=data, json_data=json_data)
+
+        original_validate_certs = self.module.params.get("validate_certs")
+        self.module.params["validate_certs"] = self.module.params.get(
+            "endpoint_tls", True
+        )
+
+        last_error = None
+        try:
+            for attempt in range(max_retries):
+                try:
+                    resp, info = fetch_url(
+                        self.module,
+                        url,
+                        data=body,
+                        headers=headers,
+                        method=method,
+                        timeout=self.timeout,
+                    )
+
+                    status_code = info.get("status", -1)
+
+                    self._capture_cookies(info)
+
+                    if status_code in squelch:
+                        return squelch[status_code]
+
+                    if status_code == -1:
+                        error_msg = info.get("msg", "Connection error")
+                        if attempt < max_retries - 1:
+                            time.sleep(compute_backoff(attempt))
+                            last_error = CdpError(
+                                f"{error_msg} for {url}", status=status_code
+                            )
+                            continue
+                        raise CdpError(f"{error_msg} for {url}", status=status_code)
+
+                    if 200 <= status_code < 300:
+                        return parse_success_response(resp, status_code)
+
+                    error_message = parse_error_message(info, status_code)
+
+                    if is_retryable_status(status_code) and attempt < max_retries - 1:
+                        time.sleep(compute_backoff(attempt))
+                        last_error = CdpError(
+                            f"{error_message} for {url}", status=status_code
+                        )
+                        continue
+
+                    raise CdpError(f"{error_message} [{status_code}] for {url}")
+
+                except CdpError:
+                    raise
+                except Exception as e:
+                    if attempt < max_retries - 1:
+                        time.sleep(compute_backoff(attempt))
+                        last_error = CdpError(f"Connection error for {url}: {str(e)}")
+                        continue
+                    raise CdpError(
+                        f"Request failed after {max_retries} attempts for {url}: {str(e)}"
+                    )
+
+            if last_error:
+                raise last_error
+            raise CdpError(f"Request failed for {url}")
+        finally:
+            if original_validate_certs is not None:
+                self.module.params["validate_certs"] = original_validate_certs
+
+    def get(self, path: str, params: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
+        """Execute an HTTP GET request."""
+        return self._make_request("GET", path, params=params)
+
+    def post(
+        self,
+        path: str,
+        data: Optional[Dict[str, Any]] = None,
+        json_data: Optional[Dict[str, Any]] = None,
+        squelch: Dict[int, Any] = {},
+    ) -> Dict[str, Any]:
+        """Execute an HTTP POST request."""
+        return self._make_request(
+            "POST", path, data=data, json_data=json_data, squelch=squelch
+        )
+
+    def put(
+        self,
+        path: str,
+        data: Optional[Dict[str, Any]] = None,
+        json_data: Optional[Dict[str, Any]] = None,
+        squelch: Dict[int, Any] = {},
+    ) -> Dict[str, Any]:
+        """Execute an HTTP PUT request."""
+        return self._make_request(
+            "PUT", path, data=data, json_data=json_data, squelch=squelch
+        )
+
+    def delete(self, path: str, squelch: Dict[int, Any] = {}) -> Dict[str, Any]:
+        """Execute an HTTP DELETE request."""
+        return self._make_request("DELETE", path, squelch=squelch)
+
+
+class DataServiceClient(WorkloadTransport):
+    """
+    CDP Data Service workload client using Bearer token authentication.
+
+    The token and base URL are discovered through the control plane (e.g. IAM
+    ``generateWorkloadAuthToken`` or ``initiateDeployment``) and supplied here.
+    It is currently used for the DataFlow workload APIs (``/dfx/api/...``) but
+    is service-agnostic. All transport behaviour comes from WorkloadTransport;
+    this class only declares the Bearer auth headers.
+    """
+
+    def __init__(
+        self,
+        module,
+        base_url: str,
+        access_token: str,
+        timeout_seconds: int = 60,
+        default_page_size: int = 100,
+    ):
+        """
+        Initialize the workload client.
+
+        Args:
+            module: AnsibleModule instance (used for ``fetch_url`` and TLS config)
+            base_url: Base URL of the workload API
+            access_token: JWT Bearer token from generateWorkloadAuthToken
+            timeout_seconds: Per-request HTTP timeout in seconds
+            default_page_size: Default page size for paginated requests
+        """
+        super().__init__(default_page_size=default_page_size)
+        self.module = module
+        self.base_url = base_url.rstrip("/")
+        self.access_token = access_token
+        self.timeout = timeout_seconds
+        self.cookies: Dict[str, str] = {}
+        self.headers = {
+            "Content-Type": "application/json",
+            "Accept": "application/json",
+            "Authorization": f"Bearer {self.access_token}",
+        }
